fix(doc-retrieval-mcp): Resolve git-diff range against origin/ in CI
getChangedFiles() built its diff range from the bare GITHUB_BASE_REF
value ("main"), but actions/checkout never creates a local branch of
that name -- only the origin/ remote-tracking ref exists, even under
fetch-depth: 0. The Retrieval Eval Gate job failed with "unknown
revision" on this PR, and Item #2's fail-closed handling correctly
turned that into a hard CI failure instead of the old silent
pass:true -- proving the fix in production.

Prefix both branches of the range with origin/, matching the same
pattern already used elsewhere in ci.yml (change classification,
migration-ordering checks) under the same fetch-depth: 0 config.
Corrected the first regression test to use a non-"main" base ref so
it actually exercises the interpolation branch instead of
coincidentally matching the fallback string.

Split the two new tests plus the existing getChangedFiles/
evaluateDriftWithDiffResult/main CLI-wiring blocks into a new sibling
file, check-baseline-drift-diffresult.test.ts, since the additions
pushed check-baseline-drift.test.ts to 512 lines (over the 500-line
gate).

Reviewed by Opus (approve with the test-baseRef fix, applied) and
Codex (approve as-is, confirming the origin/ prefix matches existing
repo convention).

Refs SMI-5708
